feat: Add Share App option in Settings
