+++ app/src/main/res/values/strings.xml
@@ -99,6 +99,7 @@ https://github.com/RocketChat/java-code-styles/blob/master/CODING_STYLE.md#strin
     <string-array name="widechat_settings_actions">
         <item name="item_preferences">Preferences</item>
         <item name="item_privacy">Privacy</item>
+        <item name="item_share_app">Share app</item>
         <item name="item_about">About</item>
         <item name="item_license">License</item>
         <item name="item_logout">Log Out</item>
